feat(@moeum/figma-componentgen-plugin): 폴더 정리 및 도메인 규칙 정의 #130
