+++ packages/figma-componentgen-plugin/src/common/style/global.css
